+++ BugsnagPerformance/Assets/BugsnagPerformance/Scripts/Internal/BugsnagPerformanceAutoStart.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 internal class BugsnagPerformanceAutoStart : MonoBehaviour
 {
-    [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.AfterAssembliesLoaded)]
+    [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.AfterSceneLoad)]
     static void OnBeforeSceneLoadRuntimeMethod()
     {
         var settings = Resources.Load<BugsnagPerformanceSettingsObject>("Bugsnag/BugsnagPerformanceSettingsObject");