+++ BugsnagPerformance/Assets/BugsnagPerformance/Scripts/Internal/MainThreadDispatchBehaviour.cs
@@ -29,8 +29,8 @@ namespace BugsnagUnityPerformance
     public class MainThreadDispatchBehaviour : MonoBehaviour
     {
 
-        private static MainThreadDispatchBehaviour _instance;       
-    
+        private static MainThreadDispatchBehaviour _instance;
+
         private static readonly Queue<Action> _executionQueue = new Queue<Action>();
 
 
@@ -38,6 +38,9 @@ namespace BugsnagUnityPerformance
         {
             if (_instance == null)
             {
+#if BUGSNAG_DEBUG
+                Logger.I("MainThreadDispatchBehaviour created");
+#endif
                 _instance = new GameObject("Bugsnag performance main thread dispatcher").AddComponent<MainThreadDispatchBehaviour>();
             }
             return _instance;