+++ BugsnagPerformance/Assets/BugsnagPerformance/Scripts/Models/ResourceModel.cs
@@ -18,27 +18,19 @@ namespace BugsnagUnityPerformance
         public void Configure(PerformanceConfiguration config)
         {
             attributes = new List<AttributeModel>
-               {
-                    new AttributeModel("deployment.environment", config.ReleaseStage),
-
-                    new AttributeModel("telemetry.sdk.name", "bugsnag.performance.unity"),
-
-                    new AttributeModel("telemetry.sdk.version", Version.VersionString),
-
-                    new AttributeModel("os.version", Environment.OSVersion.VersionString),
-
-                    new AttributeModel("device.model.identifier", SystemInfo.deviceModel),
-
-                    new AttributeModel("service.version", string.IsNullOrEmpty(config.AppVersion) ? Application.version : config.AppVersion),
-
-                    new AttributeModel("bugsnag.app.platform", GetPlatform()),
-
-                    new AttributeModel("bugsnag.runtime_versions.unity", Application.unityVersion),
-               };
-      
-            attributes.Add(GetMobileVersionInfo(config));
-            attributes.Add(GetMobileManufacturer());
-            attributes.Add(GetMobileArch());
+            {
+                new AttributeModel("deployment.environment", config.ReleaseStage),
+                new AttributeModel("telemetry.sdk.name", "bugsnag.performance.unity"),
+                new AttributeModel("telemetry.sdk.version", Version.VersionString),
+                new AttributeModel("device.model.identifier", SystemInfo.deviceModel),
+                new AttributeModel("service.version", string.IsNullOrEmpty(config.AppVersion) ? Application.version : config.AppVersion),
+                new AttributeModel("bugsnag.app.platform", GetPlatform()),
+                new AttributeModel("bugsnag.runtime_versions.unity", Application.unityVersion),
+                GetNativeVersionInfo(config),
+                GetManufacturer(),
+                GetArch(),
+                GetNativeOsVersion()
+            };
         }
 
         public void Start()
@@ -54,23 +46,27 @@ namespace BugsnagUnityPerformance
                     return "iOS";
                 case RuntimePlatform.Android:
                     return "Android";
+                case RuntimePlatform.OSXPlayer:
+                case RuntimePlatform.OSXEditor:
+                    return "MacOS";
             }
             return string.Empty;
         }
 
-        private AttributeModel GetMobileVersionInfo(PerformanceConfiguration config)
+        private AttributeModel GetNativeVersionInfo(PerformanceConfiguration config)
         {
             switch (Application.platform)
             {
                 case RuntimePlatform.IPhonePlayer:
-                    return GetiOSBundleVersion(config);
+                case RuntimePlatform.OSXPlayer:
+                    return GetCocoaBundleVersion(config);
                 case RuntimePlatform.Android:
                     return GetAndroidVersionCode(config);
             }
             return null;
         }
 
-        private AttributeModel GetiOSBundleVersion(PerformanceConfiguration config)
+        private AttributeModel GetCocoaBundleVersion(PerformanceConfiguration config)
         {
             if (!string.IsNullOrEmpty(config.BundleVersion))
             {
@@ -88,10 +84,14 @@ namespace BugsnagUnityPerformance
             return new AttributeModel("device.version_code", AndroidNative.GetVersionCode());
         }
 
-        private AttributeModel GetMobileArch()
+        private AttributeModel GetArch()
         {
             switch (Application.platform)
             {
+                
+                case RuntimePlatform.OSXPlayer:
+                case RuntimePlatform.OSXEditor:
+                    return new AttributeModel("host.arch", MacOSNative.GetArch());
                 case RuntimePlatform.IPhonePlayer:
                     return new AttributeModel("host.arch", iOSNative.GetArch());
                 case RuntimePlatform.Android:
@@ -100,16 +100,36 @@ namespace BugsnagUnityPerformance
             return null;
         }
 
-        private AttributeModel GetMobileManufacturer()
+        private AttributeModel GetManufacturer()
         {
             switch (Application.platform)
             {
+                case RuntimePlatform.OSXEditor:
+                case RuntimePlatform.OSXPlayer:
                 case RuntimePlatform.IPhonePlayer:
                     return new AttributeModel("device.manufacturer", "Apple");
                 case RuntimePlatform.Android:
                     return new AttributeModel("device.manufacturer", AndroidNative.GetManufacturer());
             }
             return null;
+        }
+
+        private AttributeModel GetNativeOsVersion()
+        {
+            string osVersion = string.Empty;
+            switch (Application.platform)
+            {
+                case RuntimePlatform.OSXPlayer:
+                    osVersion = MacOSNative.GetOsVersion();
+                    break;
+                case RuntimePlatform.IPhonePlayer:
+                    osVersion = iOSNative.GetOsVersion();
+                    break;
+                case RuntimePlatform.Android:
+                    osVersion = AndroidNative.GetOsVersion();
+                    break;
+            }
+            return new AttributeModel("os.version", osVersion);
         }
     }
 }