+++ BugsnagPerformance/Assets/BugsnagPerformance/Scripts/Native/iOSNative.cs
@@ -1,31 +1,32 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 #if UNITY_IOS && !UNITY_EDITOR
-
 using System.Runtime.InteropServices;
-
 #endif
-
 using UnityEngine;
 
 namespace BugsnagUnityPerformance
 {
     internal class iOSNative
     {
-#if UNITY_IOS && !UNITY_EDITOR
         const string Import = "__Internal";
 
+#if UNITY_IOS && !UNITY_EDITOR
         [DllImport(Import)]
-        internal static extern string bugsnag_performance_getBundleVersion();
+        internal static extern string bugsnag_unity_performance_getBundleVersion();
 
         [DllImport(Import)]
-        internal static extern string bugsnag_performance_get_arch();
+        internal static extern string bugsnag_unity_performance_get_arch();
+
+        [DllImport(Import)]
+        internal static extern string bugsnag_unity_performance_get_os_version();
+        
 #endif
 
         public static string GetBundleVersion()
         {
 #if UNITY_IOS && !UNITY_EDITOR
-            return bugsnag_performance_getBundleVersion();
+            return bugsnag_unity_performance_getBundleVersion();
 #endif
             return null;
         }
@@ -33,7 +34,15 @@ namespace BugsnagUnityPerformance
         public static string GetArch()
         {
 #if UNITY_IOS && !UNITY_EDITOR
-            return bugsnag_performance_get_arch();
+            return bugsnag_unity_performance_get_arch();
+#endif
+            return null;
+        }
+
+        public static string GetOsVersion()
+        {
+#if UNITY_IOS && !UNITY_EDITOR
+            return bugsnag_unity_performance_get_os_version();
 #endif
             return null;
         }